# src/main/java/ru/skypro/homework/dto/Ads.java
package ru.skypro.homework.dto;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

import java.util.List;

@Getter
@Builder
@AllArgsConstructor
public class Ads {
    private int count;
    private List<Ad> results;
}